Move AMQP::Address to_variant to util
