+++ libraries/amqp/include/eosio/amqp/retrying_amqp_connection.hpp
@@ -59,7 +59,3 @@ private:
 };
 
 }
-
-namespace fc {
-void to_variant(const AMQP::Address& a, fc::variant& v);
-}
