+++ libraries/amqp/include/eosio/amqp/util.hpp
@@ -1,0 +1,8 @@
+#pragma once
+
+#include <amqpcpp.h>
+#include <fc/variant.hpp>
+
+namespace fc {
+void to_variant(const AMQP::Address& a, fc::variant& v);
+}
