+++ libraries/amqp/retrying_amqp_connection.cpp
@@ -1,6 +1,6 @@
 #include <amqpcpp.h>
-#include <fc/variant.hpp>
 
+#include <eosio/amqp/util.hpp>
 #include <eosio/amqp/retrying_amqp_connection.hpp>
 
 namespace eosio {
@@ -259,19 +259,4 @@ const AMQP::Address& single_channel_retrying_amqp_connection::address() const {
 
 single_channel_retrying_amqp_connection::~single_channel_retrying_amqp_connection() = default;
 
-}
-
-namespace fc {
-void to_variant(const AMQP::Address& a, fc::variant& v) {
-   std::string str(a.secure() ? "amqps://" : "amqp://");
-   str.append(a.login().user()).append(":********").append("@");
-   str.append(a.hostname().empty() ? "localhost" : a.hostname());
-   if(a.port() != 5672)
-      str.append(":").append(std::to_string(a.port()));
-   str.append("/");
-   if (a.vhost() != "/")
-      str.append(a.vhost());
-
-   v = str;
-}
 }
